OpenWeather source

+++ Core/Entities/Weather.cs
@@ -5,7 +5,7 @@ namespace WeatherAggregator.Core.Entities
     public class Weather
     {
         public DateTime Date { get; set; }
-        public float Temperature { get; set; }
+        public double Temperature { get; set; }
         public int Cloudness { get; set; }
         public Precipitation Precipitation { get; set; }
     }
